requirements.txt
requests
beautifulsoup4
lxml
psycopg2-binary
python-dotenv
openai
pyyaml
tenacity
gspread
google-auth
playwright>=1.40.0
